requirements.txt
langchain
langgraph
langchain-community 
langchain-ollama 
duckduckgo-search
fastapi
uvicorn[standard]
gunicorn
python-dotenv